Moved Fortran helper functions from napi.h to a separate file
Fixes #388

Also moved the corresponding function definitions from napi.c
Only compile the code if one of the fortran bindings is being built
